requirements.txt

torch>=2.0.0
torchvision>=0.15
numpy>=1.24
Pillow>=9.5
tk
matplotlib>=3.7
networkx>=3.1
PyYAML>=5.3.1
flask>=2.2
pyqt5<5.13
pyqtwebengine<5.13
pytest>=7.3
pytest-mock>=3.10
